feat(ios): Resource Bundle to provide the Privacy Manifest
No data is collected in accordance to Apple's definition of collected data.

https://developer.apple.com/app-store/app-privacy-details/#data-collection

“Collect” refers to transmitting data off the device in a way that allows you and/or your third-party partners to access it for a period longer than what is necessary to service the transmitted request in real time.

There is one API that requires a declared reason to use, the NSUserDefaults API.
The code is in the base Diagnostic utility class so it's available/used regardless of modules instead, but the actual execution path is only ever invoked through the Motion module.

It uses the standard user defaults object which is private to the application. CA92.1 appears to the best reason description, which describes that the SDK reads and writes information that is only
accessible by the app itself and not to other apps or processes.